requirements.txt
autopep8==1.5.4
asgiref==3.4.1
django-ckeditor==6.1.0
django-crispy-forms==1.12.0
django-js-asset==1.2.2
django-phonenumber-field==5.2.0
dj-database-url==0.5.0
gunicorn==20.1.0
mysql-connector-python==8.0.22
pycodestyle==2.6.0
python-decouple==3.3
phonenumbers==8.12.29
Pillow==8.3.2
psycopg2==2.9.1
pytz==2021.1
sqlparse==0.4.1
toml==0.10.2
Unipath==1.1
whitenoise==5.3.0
django-heroku